+++ lib/src/constants/decoration.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+//Padding
+const double tPaddingSize = 20.0;
+
+//Color
+const Color appMainColor = Colors.purple;
+const Color custColor = Colors.amber;
+const Color deliveryColor = Color.fromARGB(255, 4, 250, 13);
+const Color ownerColor = Color.fromARGB(255, 5, 138, 246);
+const Color drawerColor = Color.fromARGB(255, 242, 57, 255);
+const Color transparentClr = Colors.transparent;
+const Color textBlackColor = Colors.black;
+const Color iconBlackColor = Colors.black;
+const Color drawerImageBorderColor = Color.fromARGB(255, 101, 2, 114);
+const Color profileLayoutColor = Color.fromARGB(255, 236, 236, 236);
