+++ lib/src/features/auth/screens/forgetPswrd/forget_pswrd_option.dart
@@ -2,6 +2,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_application_1/src/constants/text_strings.dart';
 import 'package:flutter_application_1/src/features/auth/screens/forgetPswrd/forget_pswrd_btn.dart';
+import 'package:flutter_application_1/src/features/auth/screens/forgetPswrd/forget_pswrd_mail.dart';
 import 'package:flutter_application_1/src/routing/routes_const.dart';
 
 
@@ -30,6 +31,13 @@ class ForgetPasswordScreen{
               title: emailAddrtxt,
               subTitle: resetviaEmailtxt,
               onTap: (){
+                // Navigator.push(context,
+                //   MaterialPageRoute(
+                //     builder: ((context) {
+                //       return ForgetPasswordMailScreen();
+                //     })
+                //   )
+                // );
                 Navigator.of(context).pushNamed(resetPswrdEmailRoute);
               },
             ),
