+++ lib/src/routing/routes.dart
@@ -4,7 +4,7 @@ import 'package:flutter_application_1/src/features/users/business_owner/owner_ho
 import 'package:flutter_application_1/src/features/users/business_owner/owner_profile.dart';
 import 'package:flutter_application_1/src/features/users/customer/cust_homepage.dart';
 import 'package:flutter_application_1/src/features/users/customer/profile_page/cust_profile.dart';
-import 'package:flutter_application_1/src/features/users/customer/profile_page/edit_profile.dart';
+import 'package:flutter_application_1/src/features/users/customer/profile_page/cust_edit_profile.dart';
 import 'package:flutter_application_1/src/features/users/customer/test_read_data.dart';
 import 'package:flutter_application_1/src/features/users/deliveryman/delivery_homepage.dart';
 import 'package:flutter_application_1/src/features/auth/screens/login/login_page.dart';
